% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testshiny.R
\docType{package}
\name{testshiny-package}
\alias{testshiny}
\alias{testshiny-package}
\title{testshiny: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
    Use four spaces when indenting paragraphs within the Description.
}
